Move player into features module
