+++ app/features/page/[pageId]/page.tsx
@@ -12,7 +12,7 @@ import { LANGUAGE_CODES } from '@/lib/languageCodes';
 import type { LanguageCode } from '@/lib/languageCodes';
 import { WORD_LANGUAGE_LABELS } from '@/lib/wordLanguages';
 import { useSettings } from '@/app/context/SettingsContext';
-import { useAudio } from '@/app/context/AudioContext';
+import { useAudio } from '@/app/features/player/context/AudioContext';
 import { buildAudioUrl } from '@/lib/reciters';
 import { useParams } from 'next/navigation';
 import useSWR from 'swr';
