+++ app/features/player/components/PlaybackOptionsModal.tsx
@@ -1,9 +1,9 @@
 /* eslint-disable jsx-a11y/no-noninteractive-element-interactions */
 import React, { useEffect, useState } from 'react';
 import { SlidersHorizontal, Mic2, Repeat } from 'lucide-react';
-import { useAudio } from '@/app/context/AudioContext';
+import { useAudio } from '@/app/features/player/context/AudioContext';
 import { RECITERS } from '@/lib/reciters';
-import type { RepeatOptions } from '@/app/context/AudioContext';
+import type { RepeatOptions } from '@/app/features/player/context/AudioContext';
 
 interface Props {
   open: boolean;
