+++ app/features/tafsir/[surahId]/[ayahId]/_components/TafsirVerse.tsx
@@ -9,7 +9,7 @@ import {
 } from '@/app/components/common/SvgIcons';
 import { Verse as VerseType, Translation, Word } from '@/types';
 import type { LanguageCode } from '@/lib/languageCodes';
-import { useAudio } from '@/app/context/AudioContext';
+import { useAudio } from '@/app/features/player/context/AudioContext';
 import { useSettings } from '@/app/context/SettingsContext';
 import { useState } from 'react';
 import Spinner from '@/app/components/common/Spinner';
